excel-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pandas==2.2.1
openpyxl==3.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
